# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Volunteer AI API
 * OpenAPI spec version: 0.1.0
 */
export interface Ok {
  ok: boolean;
}

export interface HealthStatus {
  status: string;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  user: "user",
  volunteer: "volunteer",
  admin: "admin",
} as const;

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  user: "user",
  volunteer: "volunteer",
} as const;

export interface RegisterBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  /** @minLength 4 */
  password: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface SessionUser {
  id: number;
  name: string;
  email: string;
  role: Role;
  hasVolunteerProfile: boolean;
}

export interface MeResponse {
  user: SessionUser | null;
}

export type RequestStatus = (typeof RequestStatus)[keyof typeof RequestStatus];

export const RequestStatus = {
  open: "open",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type Urgency = (typeof Urgency)[keyof typeof Urgency];

export const Urgency = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface HelpRequest {
  id: number;
  userId: number;
  userName: string;
  title: string;
  purpose: string;
  category: string;
  description: string;
  urgency: Urgency;
  cityArea: string;
  /** Only revealed to approved volunteer and the owner */
  location?: string | null;
  /** Only revealed to approved volunteer and the owner */
  contactPhone?: string | null;
  status: RequestStatus;
  applicationsCount: number;
  approvedVolunteerId?: number | null;
  createdAt: string;
}

export interface CreateRequestBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  purpose: string;
  /** @minLength 1 */
  category: string;
  description: string;
  urgency: Urgency;
  /** @minLength 1 */
  cityArea: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  contactPhone: string;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  withdrawn: "withdrawn",
} as const;

export interface VolunteerProfile {
  id: number;
  userId: number;
  name: string;
  email: string;
  /** e.g. "Medical Helper", "Tutor" */
  roleTitle: string;
  skills: string[];
  bio: string;
  phone: string;
  cityArea: string;
  availability: string;
  experienceYears: number;
  rating: number;
  completedHelps: number;
}

export interface ApplicationWithVolunteer {
  id: number;
  requestId: number;
  message: string;
  status: ApplicationStatus;
  createdAt: string;
  volunteer: VolunteerProfile;
}

export type HelpRequestDetail = HelpRequest & {
  applications: ApplicationWithVolunteer[];
};

export interface Application {
  id: number;
  requestId: number;
  volunteerId: number;
  message: string;
  status: ApplicationStatus;
  createdAt: string;
  decidedAt?: string | null;
}

export interface ApplicationWithRequest {
  id: number;
  message: string;
  status: ApplicationStatus;
  createdAt: string;
  request: HelpRequest;
}

export interface ApplyBody {
  requestId: number;
  /** @minLength 1 */
  message: string;
}

export type DecisionBodyDecision =
  (typeof DecisionBodyDecision)[keyof typeof DecisionBodyDecision];

export const DecisionBodyDecision = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface DecisionBody {
  decision: DecisionBodyDecision;
}

export interface UpsertVolunteerProfileBody {
  /** @minLength 1 */
  roleTitle: string;
  skills: string[];
  bio: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  cityArea: string;
  /** @minLength 1 */
  availability: string;
  /** @minimum 0 */
  experienceYears: number;
}

export interface DashboardCounts {
  total: number;
  applied: number;
  pending: number;
  accepted: number;
  rejected: number;
}

export interface UserDashboard {
  counts: DashboardCounts;
  recentRequests: HelpRequest[];
  recentApplications: ApplicationWithVolunteer[];
}

export type VolunteerDashboardCounts = {
  activeRoles: number;
  applications: number;
  accepted: number;
  rejected: number;
  availableRequests: number;
};

export interface VolunteerDashboard {
  counts: VolunteerDashboardCounts;
  recentApplications: ApplicationWithRequest[];
  availableRequests: HelpRequest[];
  profile: VolunteerProfile | null;
}

export type AdminDashboardTotals = {
  users: number;
  volunteers: number;
  requests: number;
  applications: number;
  approvedHelps: number;
};

export type AdminDashboardRequestStatusBreakdownItem = {
  status: string;
  count: number;
};

export type AdminDashboardApplicationStatusBreakdownItem = {
  status: string;
  count: number;
};

export type AdminDashboardCategoryBreakdownItem = {
  category: string;
  count: number;
};

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  request_created: "request_created",
  application_created: "application_created",
  application_decided: "application_decided",
  user_registered: "user_registered",
  volunteer_registered: "volunteer_registered",
} as const;

export interface ActivityItem {
  id: string;
  kind: ActivityItemKind;
  message: string;
  actorName: string;
  createdAt: string;
}

export interface AdminDashboard {
  totals: AdminDashboardTotals;
  requestStatusBreakdown: AdminDashboardRequestStatusBreakdownItem[];
  applicationStatusBreakdown: AdminDashboardApplicationStatusBreakdownItem[];
  categoryBreakdown: AdminDashboardCategoryBreakdownItem[];
  recentActivity: ActivityItem[];
  topVolunteers: VolunteerProfile[];
}

export interface AdminUserRow {
  id: number;
  name: string;
  email: string;
  createdAt: string;
  totalRequests: number;
  openRequests: number;
  completedRequests: number;
}

export interface AdminVolunteerRow {
  id: number;
  userId: number;
  name: string;
  email: string;
  roleTitle: string;
  cityArea: string;
  totalApplications: number;
  approved: number;
  rejected: number;
  pending: number;
}

export interface AdminUserDetail {
  user: AdminUserRow;
  requests: HelpRequest[];
}

export interface AdminVolunteerDetail {
  volunteer: AdminVolunteerRow;
  profile: VolunteerProfile;
  applications: ApplicationWithRequest[];
}

export type ListRequestsParams = {
  /**
   * mine=user's own; available=open feed for volunteers; all=admin
   */
  scope?: ListRequestsScope;
};

export type ListRequestsScope =
  (typeof ListRequestsScope)[keyof typeof ListRequestsScope];

export const ListRequestsScope = {
  mine: "mine",
  available: "available",
  all: "all",
} as const;

export type ListApplicationsParams = {
  scope?: ListApplicationsScope;
};

export type ListApplicationsScope =
  (typeof ListApplicationsScope)[keyof typeof ListApplicationsScope];

export const ListApplicationsScope = {
  mine: "mine",
  all: "all",
} as const;