# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Volunteer AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user or volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (user, volunteer, or admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications visible to current role
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [mine, all]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationWithRequest"
    post:
      operationId: applyToRequest
      tags: [applications]
      summary: Volunteer applies to a request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/decision:
    post:
      operationId: decideApplication
      tags: [applications]
      summary: User approves or rejects a volunteer application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/volunteers/{id}:
    get:
      operationId: adminGetVolunteer
      tags: [admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminVolunteerDetail"
...
components:
  schemas:
...
    Role:
      type: string
      enum: [user, volunteer, admin]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 4
        role:
          type: string
          enum: [user, volunteer]
      required: [name, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    SessionUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        hasVolunteerProfile:
          type: boolean
      required: [id, name, email, role, hasVolunteerProfile]
...
    RequestStatus:
      type: string
      enum: [open, in_progress, completed, cancelled]
    Urgency:
      type: string
      enum: [low, medium, high]
    HelpRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        title:
          type: string
        purpose:
          type: string
        category:
          type: string
        description:
          type: string
        urgency:
          $ref: "#/components/schemas/Urgency"
        cityArea:
          type: string
        location:
          type: string
          nullable: true
          description: Only revealed to approved volunteer and the owner
        contactPhone:
          type: string
          nullable: true
          description: Only revealed to approved volunteer and the owner
        status:
          $ref: "#/components/schemas/RequestStatus"
        applicationsCount:
          type: integer
        approvedVolunteerId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - title
        - purpose
        - category
        - description
        - urgency
        - cityArea
        - status
        - applicationsCount
        - createdAt
...
    ApplicationStatus:
      type: string
      enum: [pending, approved, rejected, withdrawn]
    Application:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        volunteerId:
          type: integer
        message:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        createdAt:
          type: string
        decidedAt:
          type: string
          nullable: true
      required: [id, requestId, volunteerId, message, status, createdAt]
...
    ApplicationWithRequest:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        status:
          $ref: "#/components/schemas/ApplicationStatus"
        createdAt:
          type: string
        request:
          $ref: "#/components/schemas/HelpRequest"
      required: [id, message, status, createdAt, request]
    ApplyBody:
      type: object
      properties:
        requestId:
          type: integer
        message:
          type: string
          minLength: 1
      required: [requestId, message]
    DecisionBody:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
      required: [decision]
    VolunteerProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        roleTitle:
          type: string
          description: e.g. "Medical Helper", "Tutor"
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        phone:
          type: string
        cityArea:
          type: string
        availability:
          type: string
        experienceYears:
          type: integer
        rating:
          type: number
        completedHelps:
          type: integer
      required:
        - id
        - userId
        - name
        - email
        - roleTitle
        - skills
        - bio
        - phone
        - cityArea
        - availability
        - experienceYears
        - rating
        - completedHelps
...
    AdminUserRow:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
        totalRequests:
          type: integer
        openRequests:
          type: integer
        completedRequests:
          type: integer
      required: [id, name, email, createdAt, totalRequests, openRequests, completedRequests]
    AdminVolunteerRow:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        roleTitle:
          type: string
        cityArea:
          type: string
        totalApplications:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer
      required:
        - id
        - userId
        - name
        - email
        - roleTitle
        - cityArea
        - totalApplications
        - approved
        - rejected
        - pending
    AdminUserDetail:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUserRow"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/HelpRequest"
      required: [user, requests]
    AdminVolunteerDetail:
      type: object
      properties:
        volunteer:
          $ref: "#/components/schemas/AdminVolunteerRow"
        profile:
          $ref: "#/components/schemas/VolunteerProfile"
        applications:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationWithRequest"
      required: [volunteer, profile, applications]
...